add int back

+++ utils.py
@@ -34,10 +34,10 @@ def writenewIWR(scenario, all_split_data, all_data, firstline_iwr, sow, users,
             row_data[2] = str(int(float(row_data[2]) * remaining_demand))
             # scale other months
             for j in range(len(all_split_data[i + firstline_iwr]) - 2):
-                row_data.append(str(float(all_split_data[i + firstline_iwr][j + 1]) * remaining_demand))
+                row_data.append(str(int(float(all_split_data[i + firstline_iwr][j + 1]) * remaining_demand)))
         else:
             for j in range(len(all_split_data[i + firstline_iwr]) - 2):
-                row_data.append(str(float(all_split_data[i + firstline_iwr][j + 1])))
+                row_data.append(str(int(float(all_split_data[i + firstline_iwr][j + 1]))))
         # append row of adjusted data
         new_data.append(row_data)
 
@@ -81,7 +81,7 @@ def writenewDDM(scenario, all_data_DDM, firstline_ddm, CMIP_IWR,
             irrigation_encounters[index] = +1
             for m in range(len(change)):
                 change[m] = float(sample_IWR[line_in_iwr][2 + m]) - float(CMIP_IWR[line_in_iwr][2 + m])
-                row_data[m + 2] = str(float(row_data[m + 2]) + change[m])
+                row_data[m + 2] = str(int(float(row_data[m + 2]) + change[m]))
         # append row of adjusted data
         new_data.append(row_data)
         # write new data to file
